chore: Moves WarningIconWithTooltip to own folder
